+++ src/shapes.h
@@ -12,12 +12,12 @@
 
 struct shapemag 
 {
-  double bands[5];
+  float bands[5];
 };
 
 std::istream & operator>>(std::istream & is, shapemag & in);
 
-void generate_shapes(std::vector<double>& mags, std::vector<double>& e, std::vector<double>& s,
+void generate_shapes(std::vector<float>& mags, std::vector<double>& e, std::vector<double>& s,
 		     int nelem, int vl);
 
 #endif
